+++ R/rmarkdown.R
@@ -1,7 +1,8 @@
 #' Render RMarkdown document in a fresh session
 #'
 #' @noRd
-render_rmarkdown <- function(pkg, input, output, ..., copy_images = TRUE, quiet = TRUE) {
+render_rmarkdown <- function(pkg, input, output, ..., copy_images = TRUE,
+                             quiet = TRUE, parallel = FALSE) {
 
   input_path <- path_abs(input, pkg$src_path)
   output_path <- path_abs(output, pkg$dst_path)
@@ -10,7 +11,9 @@ render_rmarkdown <- function(pkg, input, output, ..., copy_images = TRUE, quiet 
     stop("Can't find ", src_path(input), call. = FALSE)
   }
 
-  cat_line("Reading ", src_path(input))
+  if (!isTRUE(parallel)) {
+    cat_line("Reading ", src_path(input))
+  }
   digest <- file_digest(output_path)
 
   args <- list(
@@ -40,7 +43,9 @@ render_rmarkdown <- function(pkg, input, output, ..., copy_images = TRUE, quiet 
     update_html(path, tweak_rmarkdown_html, input_path = path_dir(input_path))
   }
   if (digest != file_digest(output_path)) {
-    cat_line("Writing ", dst_path(output))
+    if (!isTRUE(parallel)) {
+      cat_line("Writing ", dst_path(output))
+    }
   }
 
   # Copy over images needed by the document